Ns

See merge request martynets.aleksandr/project3_1!49

+++ src/main/java/com/amr/project/model/dto/ItemDto.java
@@ -19,8 +19,8 @@ import java.util.List;
 public class ItemDto {
     private Long id;
     private String name;
-    private BigDecimal basePrice;
-    private BigDecimal price;
+    private Long basePrice;
+    private Long price;
     private int count;
     private double rating;
     private String description;
